cleanup recipes view-container separation
